Separation of Custom Game World from Normal Game World
- Added 5 stages, this is what we should be focused on.
- Added a reusable UI Overlay module, just drag and drop onto a stage and it will figure out the rest!
- Added Package Prefab folder with Exit to Menu Dialog Canvas.  Fully functional, press Escape on a screen that you've dropped this onto and options to return to menu are presented.
- Reogranized folders for prefabs, sections include Tiles (Basic movement, start/end tiles), interactables (things that affect the ball), and Pickups (things the ball uses).
- Removed world.cs
- Added two game manager's, GameManager and CustomGameManager
- Main Menu has buttons to start each of these.

Levels should be designed as standalone scenes rather than using the custom as there would need to be much more refinement on this.  Easier to just make scenes are needed.  Perhaps if we're bored after the project is done, I would implement this just to see how effective I could make a custom game editor.