Web GL build has been created
Created a Web GL build under the Game / WebBuilds folder.

All files and folders within should be copied into the folder with our website for the application.

There is an issue with loading resources that I need to look into, as it cannot find resources built or included in the build folder.  Will mess with this throughout the week.